+++ packages/db/src/schema/study-rooms.ts
@@ -1,0 +1,50 @@
+import {
+  boolean,
+  index,
+  integer,
+  pgTable,
+  timestamp,
+  uniqueIndex,
+  uuid,
+  varchar,
+} from "drizzle-orm/pg-core";
+
+export const studyLocation = pgTable("study_location", {
+  id: varchar("id").primaryKey(),
+  name: varchar("name").notNull(),
+});
+
+export const studyRoom = pgTable(
+  "study_room",
+  {
+    id: varchar("id").primaryKey(),
+    name: varchar("name").notNull(),
+    capacity: integer("capacity"),
+    location: varchar("location").notNull(),
+    description: varchar("description").notNull(),
+    directions: varchar("directions").notNull(),
+    techEnhanced: boolean("tech_enhanced"),
+    url: varchar("url"),
+    studyLocationId: varchar("study_location_id")
+      .references(() => studyLocation.id)
+      .notNull(),
+  },
+  (table) => [index().on(table.studyLocationId)],
+);
+
+export const studyRoomSlot = pgTable(
+  "study_room_slot",
+  {
+    id: uuid("id").primaryKey().defaultRandom(),
+    studyRoomId: varchar("study_room_id")
+      .references(() => studyRoom.id)
+      .notNull(),
+    start: timestamp("start", { mode: "date" }).notNull(),
+    end: timestamp("end", { mode: "date" }).notNull(),
+    isAvailable: boolean("is_available").notNull(),
+  },
+  (table) => [
+    index().on(table.studyRoomId),
+    uniqueIndex().on(table.studyRoomId, table.start, table.end),
+  ],
+);
